+++ XADMaster/XADTarParser.m
@@ -278,7 +278,7 @@
 		// Parse the pair into key/value.
 		read_length = 0;
 		start_pos = position;
-		int max_len = strlen( key_val_pair );
+		int max_len = (int)strlen( key_val_pair );
 		while( key_val_pair[read_length] != '=' && read_length < max_len ) {
 			position++;
 			read_length++;
